hotfix: Temporarily disable web session link
Temporarily disables the link to the web session viewer page from the user summary page. This is an emergency hotfix to immediately stop high-volume traffic that is causing database-wide lock contention and preventing migrations.

The link is disabled by removing the href and marking the button as disabled. The proper fix (implementing pagination on the target page) will be deployed subsequently.